Cleanup auth
