+++ services/api-gateway/src/simcore_service_api_gateway/auth.py
@@ -1,7 +1,18 @@
-#
-# FROM https://fastapi.tiangolo.com/advanced/security/oauth2-scopes/
-#
+import logging
+from typing import Optional
 
+from fastapi import Depends, HTTPException, Security, status
+from fastapi.security import OAuth2PasswordBearer, SecurityScopes
+
+from . import crud_users as crud
+from .auth_security import get_access_token_data
+from .schemas import TokenData, User, UserInDB
+
+
+log = logging.getLogger(__name__)
+
+
+# Resource SERVER ----------------------------------------------
 #
 #  +--------+                               +---------------+
 #  |        |--(A)- Authorization Request ->|   Resource    |
@@ -23,58 +34,6 @@
 #
 #                  Figure 1: Abstract Protocol Flow
 
-
-import logging
-from typing import Optional
-
-from fastapi import APIRouter, Depends, HTTPException, Security, status
-from fastapi.security import (
-    OAuth2PasswordBearer,
-    OAuth2PasswordRequestForm,
-    SecurityScopes,
-)
-
-from . import crud_users as crud
-from .auth_security import authenticate_user, create_access_token, get_access_token_data
-from .schemas import Token, TokenData, User, UserInDB
-from .utils.helpers import json_dumps
-
-log = logging.getLogger(__name__)
-
-
-# Authorization SERVER -------------------------------------
-router = APIRouter()
-
-
-@router.post("/token", response_model=Token)
-async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
-
-    print("Form Request", "-" * 20)
-    for (
-        attr
-    ) in "grant_type   username    password    scopes    client_id    client_secret".split():
-        print("-", attr, ":", getattr(form_data, attr))
-    print("-" * 20)
-
-    user = authenticate_user(form_data.username, form_data.password)
-    if not user:
-        raise HTTPException(status_code=400, detail="Incorrect username or password")
-
-    # TODO: grant requested scopes OR NOT!
-
-    access_token = create_access_token(
-        subject={"sub": user.username, "scopes": form_data.scopes}
-    )
-    #
-    resp_data = {"access_token": access_token, "token_type": "bearer"}
-
-    print("{:-^30}".format("/token response"))
-    print(json_dumps(resp_data))
-    print("-" * 30)
-    return resp_data
-
-
-# Resource SERVER ----------------------------------------------
 
 # callable with request as argument -> extracts token from Authentication header
 oauth2_scheme = OAuth2PasswordBearer(
