+++ services/api-gateway/src/simcore_service_api_gateway/auth_security.py
@@ -12,13 +12,15 @@ from jwt import PyJWTError
 from passlib.context import CryptContext
 
 from . import crud_users as crud
-from .schemas import TokenData, ValidationError, UserInDB
+from .schemas import TokenData, UserInDB, ValidationError
 
 # PASSWORDS
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
+
 def verify_password(plain_password: str, hashed_password: str) -> bool:
     return pwd_context.verify(plain_password, hashed_password)
+
 
 def get_password_hash(password: str) -> str:
     return pwd_context.hash(password)
@@ -42,15 +44,16 @@ def create_secret_key() -> str:
     return str(proc.stdout).strip()
 
 
+# JSON WEB TOKENS (JWT)
+
 __SIGNING_KEY__ = os.environ.get("SECRET_KEY") or create_secret_key()
-__ALGORITHM__  = "HS256"
+__ALGORITHM__ = "HS256"
 ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
 
-def create_access_token(*,
-    subject: str,
-    scopes: List[str]=None,
-    expires_delta: timedelta = None) -> str:
+def create_access_token(
+    *, subject: str, scopes: List[str] = None, expires_delta: timedelta = None
+) -> str:
     if expires_delta is None:
         expires_delta = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
 
@@ -58,7 +61,7 @@ def create_access_token(*,
     to_encode = {
         "sub": subject,
         "exp": datetime.utcnow() + expires_delta,
-        "scopes": scopes or []
+        "scopes": scopes or [],
     }
     encoded_jwt = jwt.encode(to_encode, __SIGNING_KEY__, algorithm=__ALGORITHM__)
     return encoded_jwt
@@ -68,16 +71,14 @@ def decode_token(encoded_jwt: str) -> dict:
     return jwt.decode(encoded_jwt, __SIGNING_KEY__, algorithms=[__ALGORITHM__])
 
 
-
-
-def get_access_token_data(token: str) -> Optional[TokenData]:
+def get_access_token_data(encoded_jwt: str) -> Optional[TokenData]:
     """
         Decodes and validates
     """
     # returns valid
     try:
         # decode JWT [header.payload.signature] and get payload:
-        payload = decode_token(token)
+        payload = decode_token(encoded_jwt)
 
         username: str = payload.get("sub")
         if username is None:
